Unmap identity map after jump to higher half
We also have to modify the multiboot info's addresses to point into our
higher half, which I've done with a simple little macro in assembly.

It would probably almost certainly be nicer to to so from within C,
but I like the idea that everything's just set up by that point.
(Well except gdt I suppose ;)