Only map .data/.bss sections as writable
This took so long because I kept shooting myself in the foot.
Funny how that works isn't it?

When moving the MB adjustment code up, I didn't move the adjustment to
the pointer itself, so I was then trying to write into some memory that
clearly doesn't exist, which caused a triple fault..

I also just completely forgot the moving code, so a test I had in
`higher_half` was trying to write into some readonly memory
(`0x0`, of course) which *also* caused a triple fault.